requirements.txt
django
robotframework
robotframework-seleniumlibrary